attended gyro comp, no succes

+++ display.h
@@ -4,6 +4,6 @@
 void setupDisplay();
 //displays 4 integer (distances)
 void displayDistances(int distances[4]);
-void displayYPR(float yaw, float pitch, float roll);
+void displayYPR(int yaw, int pitch, int roll);
 void displayMessage(const char * msg);
